Implement an outline for more different CSV types to support for the native CSV parser

+++ src/main/resources/com/nec/arrow/functions/cpp/transfer-definitions.h
@@ -30,7 +30,7 @@ typedef struct
 
 typedef struct
 {
-    long *data;
+    long long *data;
     int count;
 } non_null_bigint_vector;
